--- __tests__/source.spec.ts
@@ -1,3 +1,0 @@
-import { IndexedStringSource } from "../source"
-
-// まだ
